[Rename] WORKSPACE_006, WORKSPACE_010 - 프로젝트 구조 변경으로 인한 파일 이동
