+++ backend/src/main/java/minionz/backend/scrum/label_select/model/SprintLabelSelect.java
@@ -1,9 +1,9 @@
-package minionz.backend.scrum.sprint_label_select.model;
+package minionz.backend.scrum.label_select.model;
 
 import jakarta.persistence.*;
 import lombok.*;
-import minionz.backend.scrum.sprint.model.Sprint;
 import minionz.backend.scrum.label.model.SprintLabel;
+import minionz.backend.scrum.sprint.model.Sprint;
 
 @Getter
 @Setter
