+++ backend/src/main/java/minionz/backend/scrum/sprint/model/Sprint.java
@@ -3,9 +3,10 @@ package minionz.backend.scrum.sprint.model;
 import jakarta.persistence.*;
 import lombok.*;
 import minionz.backend.common.BaseEntity;
+import minionz.backend.scrum.label_select.model.SprintLabelSelect;
 import minionz.backend.scrum.task.model.Task;
 import minionz.backend.scrum.meeting.model.Meeting;
-import minionz.backend.scrum.sprint_label_select.model.SprintLabelSelect;
+
 import minionz.backend.scrum.sprint_participation.model.SprintParticipation;
 import minionz.backend.scrum.workspace.model.Workspace;
