+++ backend/src/main/java/minionz/backend/scrum/task/model/Task.java
@@ -3,7 +3,7 @@ package minionz.backend.scrum.task.model;
 import jakarta.persistence.*;
 import lombok.*;
 import minionz.backend.common.BaseEntity;
-import minionz.backend.scrum.task_label_select.model.TaskLabelSelect;
+import minionz.backend.scrum.label_select.model.TaskLabelSelect;
 import minionz.backend.scrum.task_participation.model.TaskParticipation;
 import minionz.backend.scrum.meeting.model.Meeting;
 import minionz.backend.scrum.sprint.model.Sprint;
